pyproject.toml
[project]
name = "sofa-recommendation-demo"
version = "0.1.0"
description = "OceanBase Multi-Model Sofa Recommendation Demo"
authors = [
    {name = "Your Name", email = "your.email@example.com"}
]
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "pyobvector>=0.2.6",
    "sqlalchemy>=2.0.0",
    "dashscope>=1.20.12",
    "requests>=2.32.3",
    "pandas>=2.2.3",
    "tqdm>=4.66.5",
    "python-dotenv>=1.0.1",
    "pillow>=10.4.0",
    "click>=8.1.7",
    "pydantic>=2.0.0",
    "langgraph>=0.6.6",
    "streamlit>=1.30.0",
]

[tool.hatch.build.targets.wheel]
packages = ["srd"]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
